requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0
keras==2.13.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
bokeh==3.2.1

# Deep Learning & Clustering
umap-learn==0.5.3
hdbscan==0.8.29
spectral-clustering==0.2.0
sklearn-extra==0.3.0
pyclustering==0.10.1.2

# Azure Services
azure-storage-blob==12.17.0
azure-cosmos==4.5.1
azure-servicebus==7.11.0
azure-functions==1.18.0
azure-identity==1.14.0
azure-keyvault-secrets==4.7.0
azure-monitor-query==1.2.0

# API Framework
fastapi==0.101.1
uvicorn[standard]==0.23.2
pydantic==2.1.1
httpx==0.24.1
python-multipart==0.0.6

# Data Processing
dask[complete]==2023.7.1
polars==0.18.15
pyarrow==12.0.1
h5py==3.9.0

# Machine Learning Operations
mlflow==2.5.0
wandb==0.15.8
tensorboard==2.13.0
optuna==3.2.0

# Utilities
python-dotenv==1.0.0
click==8.1.6
rich==13.5.2
typer==0.9.0
pyyaml==6.0.1
toml==0.10.2
requests==2.31.0

# Development & Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.7.0
isort==5.12.0
flake8==6.0.0
mypy==1.5.1

# Jupyter & Visualization
jupyter==1.0.0
jupyterlab==4.0.5
ipywidgets==8.1.0
nbconvert==7.7.3

# Monitoring & Logging
prometheus-client==0.17.1
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
